feat(oss): sync editor/preview theme overlay + pane bg/text; lint stability in vite config; fix pro-flag imports for OSS\n\n- Add CSS-vars overlay in CodeMirrorEditor to unify bg/text\n- Add bg/text to SplitPane panes\n- Add /* eslint-env node */ to vite.config.ts\n- Add utils/proFlags.ts and point imports away from ../pro
