+++ src/components/SettingsPanel.tsx
@@ -2,7 +2,7 @@ import React, { useMemo, useState } from 'react'
 import { AppSettings, ThemeMode } from '../types/settings'
 import ShortcutEditor from './ShortcutEditor'
 import './SettingsPanel.css'
-import { isProBuild } from '../pro'
+import { isProBuild } from '../utils/proFlags'
 
 interface SettingsPanelProps {
   isOpen: boolean
